+++ inst/www/shared/shiny.js
@@ -2,6 +2,8 @@
 
 var _typeof = typeof Symbol === "function" && typeof Symbol.iterator === "symbol" ? function (obj) { return typeof obj; } : function (obj) { return obj && typeof Symbol === "function" && obj.constructor === Symbol && obj !== Symbol.prototype ? "symbol" : typeof obj; };
 
+var _slicedToArray = function () { function sliceIterator(arr, i) { var _arr = []; var _n = true; var _d = false; var _e = undefined; try { for (var _i = arr[Symbol.iterator](), _s; !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i["return"]) _i["return"](); } finally { if (_d) throw _e; } } return _arr; } return function (arr, i) { if (Array.isArray(arr)) { return arr; } else if (Symbol.iterator in Object(arr)) { return sliceIterator(arr, i); } else { throw new TypeError("Invalid attempt to destructure non-iterable instance"); } }; }();
+
 function _defineProperty(obj, key, value) { if (key in obj) { Object.defineProperty(obj, key, { value: value, enumerable: true, configurable: true, writable: true }); } else { obj[key] = value; } return obj; }
 
 //---------------------------------------------------------------------
@@ -247,6 +249,261 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
       if (obj.hasOwnProperty(key)) newObj[key] = f(obj[key]);
     }
     return newObj;
+  }
+
+  // Binary equality function used by the equal function.
+  function _equal(x, y) {
+    if ($.type(x) === "object" && $.type(y) === "object") {
+      if (Object.keys(x).length !== Object.keys(y).length) return false;
+      for (var prop in x) {
+        if (!y.hasOwnProperty(prop) || !_equal(x[prop], y[prop])) return false;
+      }return true;
+    } else if ($.type(x) === "array" && $.type(y) === "array") {
+      if (x.length !== y.length) return false;
+      for (var i = 0; i < x.length; i++) {
+        if (!_equal(x[i], y[i])) return false;
+      }return true;
+    } else {
+      return x === y;
+    }
+  }
+
+  // Structural or "deep" equality predicate. Tests two or more arguments for
+  // equality, traversing arrays and objects (as determined by $.type) as
+  // necessary.
+  //
+  // Objects other than objects and arrays are tested for equality using ===.
+  function equal() {
+    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {
+      args[_key] = arguments[_key];
+    }
+
+    if (args.length < 2) throw new Error("equal requires at least two arguments.");
+    for (var i = 0; i < args.length - 1; i++) {
+      if (!_equal(args[i], args[i + 1])) return false;
+    }
+    return true;
+  };
+
+  // multimethod: Creates functions — "multimethods" — that are polymorphic on one
+  // or more of their arguments.
+  //
+  // Multimethods can take any number of arguments. Arguments are passed to an
+  // applicable function or "method", returning its result. By default, if no
+  // method was applicable, an exception is thrown.
+  //
+  // Methods are searched in the order that they were added, and the first
+  // applicable method found is the one used.
+  //
+  // A method is applicable when the "dispatch value" associated with it
+  // corresponds to the value returned by the dispatch function. The dispatch
+  // function defaults to the value of the first argument passed to the
+  // multimethod.
+  //
+  // The correspondence between the value returned by the dispatch function and
+  // any method's dispatch value is determined by the test function, which is
+  // user-definable and defaults to `equal` or deep equality.
+  //
+  // # Chainable Functions
+  //
+  // The function returned by `multimethod()` exposes functions as properties.
+  // These functions generally return the multimethod, and so can be chained.
+  //
+  // - dispatch([function newDispatch]): Sets the dispatch function. The dispatch
+  //   function can take any number of arguments, but must return a dispatch
+  //   value. The default dispatch function returns the first argument passed to
+  //   the multimethod.
+  //
+  // - test([function newTest]): Sets the test function. The test function takes
+  //   two arguments: the dispatch value produced by the dispatch function, and
+  //   the dispatch value associated with some method. It must return a boolean
+  //   indicating whether or not to select the method. The default test function
+  //   is `equal`.
+  //
+  // - when(object dispatchVal, function method): Adds a new dispatch value/method
+  //   combination.
+  //
+  // - whenAny(array<object> dispatchVals, function method): Like `when`, but
+  //   associates the method with every dispatch value in the `dispatchVals`
+  //   array.
+  //
+  // - else(function newDefaultMethod): Sets the default function. This function
+  //   is invoked when no methods apply. If left unset, the multimethod will throw
+  //   an exception when no methods are applicable.
+  //
+  // - clone(): Returns a new, functionally-equivalent multimethod. This is a way
+  //   to extend an existing multimethod in a local context — such as inside a
+  //   function — without modifying the original. NOTE: The array of methods is
+  //   copied, but the dispatch values themselves are not.
+  //
+  // # Self-reference
+  //
+  // The multimethod function can be obtained inside its method bodies without
+  // referring to it by name.
+  //
+  // This makes it possible for one method to call another, or to pass the
+  // multimethod to other functions as a callback from within methods.
+  //
+  // The mechanism is: the multimethod itself is bound as `this` to methods when
+  // they are called. Since arrow functions cannot be bound to objects, **self-reference
+  // is only possible within methods created using the `function` keyword**.
+  //
+  // # Tail recursion
+  //
+  // A method can call itself in a way that will not overflow the stack by using
+  // `this.recur`.
+  //
+  // `this.recur` is a function available in methods created using `function`.
+  // When the return value of a call to `this.recur` is returned by a method, the
+  // arguments that were supplied to `this.recur` are used to call the
+  // multimethod.
+  //
+  // # Examples
+  //
+  // Handling events:
+  //
+  //    var handle = multimethod()
+  //     .dispatch(e => [e.target.tagName.toLowerCase(), e.type])
+  //     .when(["h1", "click"], e => "you clicked on an h1")
+  //     .when(["p", "mouseover"], e => "you moused over a p"})
+  //     .else(e => {
+  //       let tag = e.target.tagName.toLowerCase();
+  //       return `you did ${e.type} to an ${tag}`;
+  //     });
+  //
+  //    $(document).on("click mouseover mouseup mousedown", e => console.log(handle(e)))
+  //
+  // Self-calls:
+  //
+  //    var demoSelfCall = multimethod()
+  //     .when(0, function(n) {
+  //       this(1);
+  //     })
+  //     .when(1, function(n) {
+  //       doSomething(this);
+  //     })
+  //     .when(2, _ => console.log("tada"));
+  //
+  // Using (abusing?) the test function:
+  //
+  //    var fizzBuzz = multimethod()
+  //     .test((x, divs) => divs.map(d => x % d === 0).every(Boolean))
+  //     .when([3, 5], x => "FizzBuzz")
+  //     .when([3], x => "Fizz")
+  //     .when([5], x => "Buzz")
+  //     .else(x => x);
+  //
+  //    for(let i = 0; i <= 100; i++) console.log(fizzBuzz(i));
+  //
+  // Getting carried away with tail recursion:
+  //
+  //    var factorial = multimethod()
+  //     .when(0, () => 1)
+  //     .when(1, (_, prod = 1) => prod)
+  //     .else(function(n, prod = 1) {
+  //       return this.recur(n-1, n*prod);
+  //     });
+  //
+  //    var fibonacci = multimethod()
+  //     .when(0, (_, a = 0) => a)
+  //     .else(function(n, a = 0, b = 1) {
+  //       return this.recur(n-1, b, a+b);
+  //     });
+  function multimethod() {
+    var dispatch = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : function (firstArg) {
+      return firstArg;
+    };
+    var test = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : equal;
+    var defaultMethod = arguments.length > 2 && arguments[2] !== undefined ? arguments[2] : null;
+    var methods = arguments.length > 3 && arguments[3] !== undefined ? arguments[3] : [];
+
+
+    var trampolining = false;
+
+    function Sentinel(args) {
+      this.args = args;
+    }
+
+    function trampoline(f) {
+      return function () {
+        for (var _len2 = arguments.length, args = Array(_len2), _key2 = 0; _key2 < _len2; _key2++) {
+          args[_key2] = arguments[_key2];
+        }
+
+        trampolining = true;
+        var ret = f.apply(invoke, args);
+        while (ret instanceof Sentinel) {
+          ret = f.apply(invoke, ret.args);
+        }trampolining = false;
+        return ret;
+      };
+    }
+
+    var invoke = trampoline(function () {
+      for (var _len3 = arguments.length, args = Array(_len3), _key3 = 0; _key3 < _len3; _key3++) {
+        args[_key3] = arguments[_key3];
+      }
+
+      var dispatchVal = dispatch.apply(null, args);
+      for (var i = 0; i < methods.length; i++) {
+        var _methods$i = _slicedToArray(methods[i], 2);
+
+        var methodVal = _methods$i[0];
+        var methodFn = _methods$i[1];
+
+        if (test(dispatchVal, methodVal)) {
+          return methodFn.apply(invoke, args);
+        }
+      }
+      if (defaultMethod) {
+        return defaultMethod.apply(invoke, args);
+      } else {
+        throw new Error('No method for dispatch value ' + dispatchVal);
+      }
+    });
+
+    invoke.recur = function () {
+      for (var _len4 = arguments.length, args = Array(_len4), _key4 = 0; _key4 < _len4; _key4++) {
+        args[_key4] = arguments[_key4];
+      }
+
+      if (!trampolining) throw new Error("recur can only be called inside a method");
+      return new Sentinel(args);
+    };
+
+    invoke.dispatch = function (newDispatch) {
+      dispatch = newDispatch;
+      return invoke;
+    };
+
+    invoke.test = function (newTest) {
+      test = newTest;
+      return invoke;
+    };
+
+    invoke.when = function (dispatchVal, methodFn) {
+      methods = methods.concat([[dispatchVal, methodFn]]);
+      return invoke;
+    };
+
+    invoke.whenAny = function (dispatchVals, methodFn) {
+      return dispatchVals.reduce(function (self, val) {
+        return invoke.when(val, methodFn);
+      }, invoke);
+    };
+
+    invoke.else = function () {
+      var newDefaultMethod = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : null;
+
+      defaultMethod = newDefaultMethod;
+      return invoke;
+    };
+
+    invoke.clone = function () {
+      return multimethod(dispatch, test, defaultMethod, methods.slice());
+    };
+
+    return invoke;
   }
 
   //---------------------------------------------------------------------
@@ -5325,11 +5582,43 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
     };
   }).call(FileUploader.prototype);
 
-  function uploadFiles(evt) {
-    // If previously selected files are uploading, abort that.
-    var $el = $(evt.target);
+  // NOTE On Safari, at least version 10.1.2, *if the developer console is open*,
+  // setting the input's value will behave strangely because of a Safari bug. The
+  // uploaded file's name will appear over the placeholder value, instead of
+  // replacing it. The workaround is to restart Safari. When I (Alan Dipert) ran
+  // into this bug Winston Chang helped me diagnose the exact problem, and Winston
+  // then submitted a bug report to Apple.
+  function setFileText($el, files) {
+    var $fileText = $el.closest('div.input-group').find('input[type=text]');
+    if (files.length === 1) {
+      $fileText.val(files[0].name);
+    } else {
+      $fileText.val(files.length + " files");
+    }
+  }
+
+  // If previously selected files are uploading, abort that.
+  function abortCurrentUpload($el) {
     var uploader = $el.data('currentUploader');
     if (uploader) uploader.abort();
+    // Clear data-restore attribute if present.
+    $el.removeAttr('data-restore');
+  }
+
+  function uploadDroppedFilesIE10Plus(el, files) {
+    var $el = $(el);
+    abortCurrentUpload($el);
+
+    // Set the label in the text box
+    setFileText($el, files);
+
+    // Start the new upload and put the uploader in 'currentUploader'.
+    $el.data('currentUploader', new FileUploader(exports.shinyapp, fileInputBinding.getId(el), files, el));
+  }
+
+  function uploadFiles(evt) {
+    var $el = $(evt.target);
+    abortCurrentUpload($el);
 
     var files = evt.target.files;
     // IE8 here does not necessarily mean literally IE8; it indicates if the web
@@ -5339,18 +5628,13 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
 
     if (!IE8 && files.length === 0) return;
 
-    // Clear data-restore attribute if present.
-    $el.removeAttr('data-restore');
-
     // Set the label in the text box
     var $fileText = $el.closest('div.input-group').find('input[type=text]');
     if (IE8) {
       // If we're using IE8/9, just use this placeholder
       $fileText.val("[Uploaded file]");
-    } else if (files.length === 1) {
-      $fileText.val(files[0].name);
     } else {
-      $fileText.val(files.length + " files");
+      setFileText($el, files);
     }
 
     // Start the new upload and put the uploader in 'currentUploader'.
@@ -5361,6 +5645,12 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
       $el.data('currentUploader', new FileUploader(exports.shinyapp, id, files, evt.target));
     }
   }
+
+  // Here we maintain a list of all the current file inputs. This is necessary
+  // because we need to trigger events on them in order to respond to file drag
+  // events. For example, they should all light up when a file is dragged on to
+  // the page.
+  var $fileInputs = $();
 
   var fileInputBinding = new InputBinding();
   $.extend(fileInputBinding, {
@@ -5406,11 +5696,206 @@ function _defineProperty(obj, key, value) { if (key in obj) { Object.definePrope
       // This will be used only when restoring a file from a saved state.
       return 'shiny.file';
     },
-    subscribe: function subscribe(el, callback) {
-      $(el).on('change.fileInputBinding', uploadFiles);
+    _getZone: function _getZone(el) {
+      return $(el).closest("div.input-group");
     },
+    // This implements draghoverstart/draghoverend events that occur once per
+    // selector, instead of once for every child the way native
+    // dragenter/dragleave do. Inspired by https://gist.github.com/meleyal/3794126
+    _enableDraghover: function _enableDraghover($el) {
+      var ns = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : "";
+
+      // Create an empty jQuery collection. This is a set-like data structure that
+      // jQuery normally uses to contain the results of a selection.
+      var collection = $();
+
+      // Attach a dragenter handler to $el and all of its children. When the first
+      // child is entered, trigger a draghoverstart event.
+      $el.on("dragenter.dragHover", function (e) {
+        if (collection.size() === 0) {
+          $el.trigger("draghoverstart" + ns, e.originalEvent);
+        }
+        // Every child that has fired dragenter is added to the collection.
+        // Addition is idempotent, which accounts for elements producing dragenter
+        // multiple times.
+        collection = collection.add(e.originalEvent.target);
+      });
+
+      // Attach dragleave and drop handlers to $el and its children. Whenever a
+      // child fires either of these events, remove it from the collection.
+      $el.on("dragleave.dragHover drop.dragHover", function (e) {
+        collection = collection.not(e.originalEvent.target);
+        // When the collection has no elements, all of the children have been
+        // removed, and produce draghoverend event.
+        if (collection.size() === 0) {
+          $el.trigger("draghoverend" + ns, e.originalEvent);
+        }
+      });
+    },
+    _disableDraghover: function _disableDraghover($el) {
+      $el.off(".dragHover");
+    },
+    _enableDocumentEvents: function _enableDocumentEvents() {
+      var $doc = $("html");
+
+      this._enableDraghover($doc);
+      $doc.on({
+        "draghoverstart.fileDrag": function draghoverstartFileDrag(e) {
+          $fileInputs.trigger("showZone.fileDrag");
+        },
+        "draghoverend.fileDrag": function draghoverendFileDrag(e) {
+          $fileInputs.trigger("hideZone.fileDrag");
+        },
+        "dragover.fileDrag drop.fileDrag": function dragoverFileDragDropFileDrag(e) {
+          e.preventDefault();
+        }
+      });
+    },
+    _disableDocumentEvents: function _disableDocumentEvents() {
+      var $doc = $("html");
+
+      $doc.off(".fileDrag");
+      this._disableDraghover($doc);
+    },
+    _zoneEvents: ["showZone.fileDrag", "hideZone.fileDrag", "draghoverstart.zone", "draghoverend.zone", "drop"].join(" "),
+    _canSetFiles: function _canSetFiles(fileList) {
+      var testEl = document.createElement("input");
+      testEl.type = "file";
+      try {
+        testEl.files = fileList;
+      } catch (e) {
+        return false;
+      }
+      return true;
+    },
+    _handleDrop: function _handleDrop(e, el) {
+      var files = e.originalEvent.dataTransfer.files,
+          $el = $(el);
+      if (files === undefined || files === null) {
+        // 1. The FileList object isn't supported by this browser, and
+        // there's nothing else we can try. (< IE 10)
+        console.log("Dropping files is not supported on this browser. (no FileList)");
+      } else if (!this._canSetFiles(files)) {
+        // 2. The browser doesn't support assigning a type=file input's .files
+        // property, but we do have a FileList to work with. (IE10+/Edge)
+        $el.val("");
+        uploadDroppedFilesIE10Plus(el, files);
+      } else {
+        // 3. The browser supports FileList and input.files assignment.
+        // (Chrome, Safari)
+        $el.val("");
+        el.files = e.originalEvent.dataTransfer.files;
+      }
+    },
+    _activeClass: "shiny-file-input-active",
+    _overClass: "shiny-file-input-over",
+    _isIE9: function _isIE9() {
+      try {
+        return window.navigator.userAgent.match(/MSIE 9\./) && true || false;
+      } catch (e) {
+        return false;
+      }
+    },
+    subscribe: function subscribe(el, callback) {
+      var _this = this;
+
+      var $el = $(el);
+      // Here we try to set up the necessary events for Drag and Drop ("DnD") on
+      // every browser except IE9. We specifically exclude IE9 because it's one
+      // browser that supports just enough of the functionality we need to be
+      // confusing. In particular, it supports drag events, so drop zones will
+      // highlight when a file is dragged into the browser window. It doesn't
+      // support the FileList object though, so the user's expectation that DnD is
+      // supported based on this highlighting would be incorrect.
+      if (!this._isIE9()) {
+        (function () {
+          var $zone = _this._getZone(el),
+              getState = function getState() {
+            return $el.data("state");
+          },
+              setState = function setState(newState) {
+            return $el.data("state", newState);
+          },
+              transition = multimethod().dispatch(function (e) {
+            return [getState(), e.type];
+          }).when(["plain", "showZone"], function (e) {
+            $zone.removeClass(_this._overClass);
+            $zone.addClass(_this._activeClass);
+            setState("activated");
+          }).when(["activated", "hideZone"], function (e) {
+            $zone.removeClass(_this._overClass);
+            $zone.removeClass(_this._activeClass);
+            setState("plain");
+          }).when(["activated", "draghoverstart"], function (e) {
+            $zone.addClass(_this._overClass);
+            $zone.removeClass(_this._activeClass);
+            setState("over");
+          })
+          // A "drop" event always coincides with a "draghoverend" event. Since
+          // we handle all draghoverend events the same way, by clearing our
+          // over-style and reverting to "activated" state, we only need to
+          // worry about handling the file upload itself here.
+          .when(["over", "drop"], function (e) {
+            _this._handleDrop(e, el);
+            // State change taken care of by ["over", "draghoverend"] handler.
+          }).when(["over", "draghoverend"], function (e) {
+            $zone.removeClass(_this._overClass);
+            $zone.addClass(_this._activeClass);
+            setState("activated");
+          })
+          // This next case happens when the window (like Finder) that a file is
+          // being dragged from occludes the browser window, and the dragged
+          // item first enters the page over a drop zone instead of entering
+          // through a none-zone element.
+          //
+          // The dragenter event that caused this draghoverstart to occur will
+          // bubble to the document, where it will cause a showZone event to be
+          // fired, and drop zones will activate and their states will
+          // transition to "activated".
+          //
+          // We schedule a function to be run *after* that happens, using
+          // setTimeout. The function we schedule will set the current element's
+          // state to "over", preparing us to deal with a subsequent
+          // "draghoverend".
+          .when(["plain", "draghoverstart"], function (e) {
+            window.setTimeout(function () {
+              $zone.addClass(_this._overClass);
+              $zone.removeClass(_this._activeClass);
+              setState("over");
+            }, 0);
+          }).else(function (e) {
+            console.log("fileInput DnD unhandled transition", getState(), e.type, e);
+          });
+
+          if ($fileInputs.length === 0) _this._enableDocumentEvents();
+          setState("plain");
+          $zone.on(_this._zoneEvents, transition);
+          $fileInputs = $fileInputs.add(el);
+          _this._enableDraghover($zone, ".zone");
+        })();
+      }
+
+      $el.on("change.fileInputBinding", uploadFiles);
+    },
+
     unsubscribe: function unsubscribe(el) {
-      $(el).off('.fileInputBinding');
+      var $el = $(el),
+          $zone = this._getZone(el);
+
+      $el.removeData("state");
+
+      $zone.removeClass(this._overClass);
+      $zone.removeClass(this._activeClass);
+
+      this._disableDraghover($zone);
+
+      // Clean up local event handlers.
+      $el.off(".fileInputBinding");
+      $zone.off(this._zoneEvents);
+
+      // Remove el from list of inputs and (maybe) clean up global event handlers.
+      $fileInputs = $fileInputs.not(el);
+      if ($fileInputs.length === 0) this._disableDocumentEvents();
     }
   });
   inputBindings.register(fileInputBinding, 'shiny.fileInputBinding');